+++ src/utils/xrLC/xrPhase_AdaptiveHT.cpp
@@ -203,7 +203,7 @@ void CBuild::xrPhase_AdaptiveHT	()
 		
 		u32 threads			= u32(-1);
 		u32 rest			= u32(-1);
-		get_intervals( CPU::ID.n_threads - 2, lc_global_data()->g_vertices().size(), threads, stride, rest );
+		get_intervals( CPU::ID.n_threads - 2, (u32)lc_global_data()->g_vertices().size(), threads, stride, rest );
 		for (u32 thID=0; thID<threads; thID++)
 			precalc_base_hemi.start	( xr_new<CPrecalcBaseHemiThread> (thID,thID*stride,thID*stride + stride ) );
 		if(rest > 0)
@@ -393,7 +393,7 @@ void CBuild::u_Tesselate(tesscb_estimator* cb_E, tesscb_face* cb_F, tesscb_verte
 	g_bUnregister				= false;
 
 	u32		counter_create		= 0;
-	u32		cnt_verts			= lc_global_data()->g_vertices().size();
+	u32 cnt_verts = (u32)lc_global_data()->g_vertices().size();
 	//u32		cnt_faces			= g_faces.size();
 	
 	for (u32 I=0; I<lc_global_data()->g_faces().size(); ++I)
